# src/main/resources/org/europa/together/configuration/acl/swagger-API.json
{
  "openapi" : "3.0.1",
  "paths" : {
...
    "/acl/1/permission" : {
      "put" : {
        "operationId" : "updatePermission",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PermissionDO"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "*/*" : { }
            }
          }
        }
      },
      "post" : {
        "operationId" : "createPermission",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PermissionDO"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "*/*" : { }
            }
          }
        }
      }
    },
    "/acl/1/permission/{id}" : {
      "get" : {
        "operationId" : "fetchPermission",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "delete" : {
        "operationId" : "deletePermission",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "*/*" : { }
            }
          }
        }
      }
    },
...
    "/acl/1/permission/list/forRole/{roleName}" : {
      "get" : {
        "operationId" : "fetchAllPermissionsOfARole",
        "parameters" : [ {
          "name" : "roleName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RolesDO" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "deleteable" : {
            "type" : "boolean"
          }
        }
      },
      "PermissionDO" : {
        "type" : "object",
        "properties" : {
          "uuid" : {
            "type" : "string"
          },
          "role" : {
            "$ref" : "#/components/schemas/RolesDO"
          },
          "resource" : {
            "$ref" : "#/components/schemas/ResourcesDO"
          },
          "read" : {
            "type" : "boolean"
          },
          "create" : {
            "type" : "boolean"
          },
          "change" : {
            "type" : "boolean"
          },
          "delete" : {
            "type" : "boolean"
          }
        }
      },
      "ResourcesDO" : {
        "type" : "object",
        "properties" : {
          "view" : {
            "type" : "string"
          },
          "deleteable" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          }
        }
      }
    }
  }
}